_examples/mock_fm/requirements.txt
nats-py==2.7.2
prometheus_client==0.20.0